# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Hybrid RAG API",
    "version": "3.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Return readiness and current index statistics.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Query",
        "description": "Execute a synchronous RAG query in FastAPI's worker thread.",
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/ingest": {
      "post": {
        "summary": "Ingest",
        "description": "Atomically ingest the configured data directory.",
        "operationId": "ingest_ingest_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/metadata": {
      "get": {
        "summary": "Metadata",
        "description": "Return index parameters and vector statistics.",
        "operationId": "metadata_metadata_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Metadata Metadata Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          }
        },
        "type": "object",
        "required": [
          "status",
          "mode",
          "total_chunks",
          "embedding_model"
        ],
        "title": "HealthResponse",
        "description": "Readiness information for orchestration probes."
      },
      "IngestResponse": {
        "properties": {
          "chunks_indexed": {
            "type": "integer",
            "title": "Chunks Indexed"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "corpus_hash": {
            "type": "string",
            "title": "Corpus Hash"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          }
        },
        "type": "object",
        "required": [
          "chunks_indexed",
          "duration_seconds",
          "corpus_hash",
          "model_name"
        ],
        "title": "IngestResponse",
        "description": "Summary of an atomic index replacement."
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 3
          },
          "temperature": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.0
          },
          "confidence_threshold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Threshold"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "description": "Parameters accepted by the query endpoint."
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "latency_seconds": {
            "type": "number",
            "title": "Latency Seconds"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "citations",
          "confidence_score",
          "mode",
          "latency_seconds",
          "sources"
        ],
        "title": "QueryResponse",
        "description": "Grounded answer and retrieval telemetry."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}